+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ## 0.1.0 2020.07.30
 Initial release
+## 0.1.1 2020.08.13
+* Add WayBackURLs seeds
+* Add '-wayback' option
+* Refine regex of emails
+* Detect bad URLs